refactor: nest per-rpc files under rpc/functions/
Mirrors @vitejs/devtools-kit conventions and leaves room for sibling
files like rpc/utils.ts next to rpc/index.ts as the rpc surface grows.
