+++ examples/files-inspector/src/rpc/functions/get-cwd.ts
