+++ examples/files-inspector/src/rpc/functions/list-files.ts
