+++ examples/files-inspector/src/rpc/index.ts
@@ -1,6 +1,6 @@
 import type { RpcDefinitionsToFunctions } from 'devframe/rpc'
-import { getCwd } from './get-cwd'
-import { listFiles } from './list-files'
+import { getCwd } from './functions/get-cwd'
+import { listFiles } from './functions/list-files'
 
 export const serverFunctions = [getCwd, listFiles] as const
