+++ examples/next-runtime-snapshot/src/rpc/functions/env.ts
