+++ examples/next-runtime-snapshot/src/rpc/functions/memory.ts
