+++ examples/next-runtime-snapshot/src/rpc/functions/system.ts
