+++ examples/streaming-chat/src/rpc/index.ts
@@ -1,7 +1,7 @@
 import type { RpcDefinitionsToFunctions } from 'devframe/rpc'
-import { clear } from './clear'
-import { demoPrompts } from './demo-prompts'
-import { send } from './send'
+import { clear } from './functions/clear'
+import { demoPrompts } from './functions/demo-prompts'
+import { send } from './functions/send'
 
 export const serverFunctions = [demoPrompts, send, clear] as const
